Reduce heavy work.
There's no reason to extract and call `geoip.New()` everytime request is
arriving. No idea what brain fart has lead into this decision in the
first hand.

Also fix the context shadowing in the error group.

This is a quick and dirty solution for now, will eventually bundle it in
the nicer encapsulation.

Signed-off-by: Ville Valkonen <weezel@users.noreply.github.com>